Update for new plugins
