+++ Actions/Enable.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace VitoDeploy\LaravelReverbPlugin;
+namespace App\Vito\Plugins\RichardAnderson\LaravelReverbPlugin\Actions;
 
 use App\Actions\Worker\CreateWorker;
 use App\Actions\Worker\ManageWorker;
