--- composer.json
@@ -1,39 +1,0 @@
-{
-    "name": "vitodeploy/laravel-reverb-plugin",
-    "description": "Laravel reverb plugin for VitoDeploy",
-    "type": "library",
-    "license": "MIT",
-    "version": "1.0.0",
-    "authors": [
-        {
-            "name": "Saeed Vaziry",
-            "email": "mr.saeedvaziry@gmail.com"
-        }
-    ],
-    "scripts": {
-        "test": "vendor/bin/phpunit"
-    },
-    "require": {
-        "php": "^8.4",
-        "illuminate/support": "^12.0",
-        "illuminate/process": "^12.0",
-        "illuminate/console": "^12.0"
-    },
-    "require-dev": {
-        "laravel/pint": "^1.10"
-    },
-    "autoload": {
-        "psr-4": {
-            "VitoDeploy\\LaravelReverbPlugin\\": "src"
-        }
-    },
-    "extra": {
-        "laravel": {
-            "providers": [
-                "VitoDeploy\\LaravelReverbPlugin\\LaravelReverbPluginServiceProvider"
-            ]
-        }
-    },
-    "minimum-stability": "stable",
-    "prefer-stable": true
-}
